libroach: rename timebound.{cc,h} to table_props.{cc,h}
Release note: None
