+++ c-deps/libroach/table_props.cc
@@ -12,7 +12,7 @@
 // implied.  See the License for the specific language governing
 // permissions and limitations under the License.
 
-#include "timebound.h"
+#include "table_props.h"
 #include <rocksdb/table_properties.h>
 #include <rocksdb/slice.h>
 #include <rocksdb/status.h>
